requirements.txt
Flask==2.1.2
Flask_Cors==3.0.10
requests==2.28.2
Pillow==9.5.0
python-escpos==2.2.0